requirements.txt
flask==2.3.3
flask-cors==3.0.10
PyJWT==2.8.0
